chore(quality): clippy --fix across the workspace
Apply all machine-applicable clippy suggestions workspace-wide (needless
borrows, collapsible ifs, redundant closures/patterns, useless conversions,
`is_empty()` over len-comparisons, etc.) across koji-db/scanner/plugins/
service + model/nominatim/macros/migration. Semantics-preserving (clippy
--fix only applies machine-verified rewrites); cargo test --workspace
exits 0. Residual manual warnings (large-variant boxing, manual trait
impls) left for a focused pedantic pass.

+++ crates/algorithms/src/bootstrap/s2.rs
@@ -174,7 +174,7 @@ impl<'a> BootstrapS2<'a> {
         }
 
         let (face, i_leaf, j_leaf, _orient) = id.face_ij_orientation();
-        let shift = (MAX_LEVEL as i32 - self.level as i32) as i32;
+        let shift = (MAX_LEVEL as i32 - self.level as i32);
 
         // Convert leaf i,j to level-L grid coords (0..2^L-1).
         let i_l = i_leaf >> shift;
@@ -229,7 +229,7 @@ pub fn cells_to_nearest_face_edges(id: CellID) -> (i32, i32) {
     const MAX_SIZE: i32 = 1 << MAX_LEVEL; // leaf resolution per axis on a face
 
     let level = id.level();
-    let size: i32 = 1 << (MAX_LEVEL - level as u64); // “width” of this cell in leaf-ij units
+    let size: i32 = 1 << (MAX_LEVEL - level); // “width” of this cell in leaf-ij units
 
     // Get face-ij for the cell center at leaf resolution.
     // Most Rust ports expose something like `to_face_ij_orientation()`.
